requirements.txt
paramiko>=3.4.0
playwright>=1.40.0
google-genai>=0.1.0